Move around signing types to avoid circular dependencies.
The new EnrichedEIP712SignTypedDataRequest introduced a circular dependency between enrichment types and the signing redux-slice.  It felt strange that many of our backend processes and services that have nothing to do with redux were reaching into a redux file for type information - so I went ahead and put SignTypedDataRequest into services/signing/types.ts
