+++ background/services/internal-ethereum-provider/index.ts
@@ -18,11 +18,8 @@ import {
 } from "../chain/utils"
 import PreferenceService from "../preferences"
 import { internalProviderPort } from "../../redux-slices/utils/contract-utils"
-import {
-  SignTypedDataRequest,
-  SignDataRequest,
-  parseSigningData,
-} from "../../redux-slices/signing"
+import { SignDataRequest, parseSigningData } from "../../redux-slices/signing"
+import { SignTypedDataRequest } from "../signing/types"
 import { hexToAscii } from "../../lib/utils"
 
 // A type representing the transaction requests that come in over JSON-RPC
